+++ client/blocks/login/index.jsx
@@ -17,7 +17,7 @@ import Notice from 'calypso/components/notice';
 import WooCommerceConnectCartHeader from 'calypso/components/woocommerce-connect-cart-header';
 import WPCloudLogo from 'calypso/components/wp-cloud-logo';
 import wooDnaConfig from 'calypso/jetpack-connect/woo-dna-config';
-import isAkismetRedirect from 'calypso/lib/akismet/is-redirect-from-akismet';
+import isAkismetRedirect from 'calypso/lib/akismet/is-akismet-redirect';
 import { preventWidows } from 'calypso/lib/formatting';
 import getGravatarOAuth2Flow from 'calypso/lib/get-gravatar-oauth2-flow';
 import { getPluginTitle, getSignupUrl } from 'calypso/lib/login';
